Module mirrord::internal_proxy

Internal proxy is accepting connection from local layers and forward it to agent while having 1:1 relationship - each layer connection is another agent connection.

This might be changed later on.

The main advantage of this design is that we remove kube logic from the layer itself, thus eliminating bugs that happen due to mix of remote env vars in our code (previously was solved using envguard which wasn’t good enough)

The proxy will either directly connect to an existing agent (currently only used for tests), or let the OperatorApi handle the connection.

Functions

connect 🔒

Connects to an agent pod depending on how LayerConfig is set-up:

connect_and_ping 🔒

Connect and send ping - this is useful when working using k8s port forward since it only creates the connection after sending the first message

create_listen_socket 🔒

Creates a listening socket using socket2 to control the backlog and manage scenarios where the proxy is under heavy load. https://github.com/metalbear-co/mirrord/issues/1716#issuecomment-1663736500 in macOS backlog is documented to be hardcoded limited to 128.

create_ping_loop 🔒

detach_io 🔒 ⚠

get_agent_connect_info 🔒

ping 🔒

Sends a ping the connection and expects a pong.

print_port 🔒

Print the port for the caller (mirrord cli execution flow) so it can pass it back to the layer instances via env var.

proxy 🔒

Main entry point for the internal proxy. It listens for inbound layer connect and forwards to agent.

redirect_fd_to_dev_null 🔒 ⚠

request_pause 🔒

Request target container pause from the connected agent.